continuing reorganization of tunneling code

+++ VevaciousPlusPlus/include/TunnelingCalculation/BounceActionTunneling/ThermalActionFitter.hpp
@@ -58,12 +58,12 @@ namespace VevaciousPlusPlus
     temperatureDifference( criticalTemperature - temperatureVector[ 0 ] );
     double fittedAction( fitCoefficients[ 0 ] );
     double termContribution( 0.0 );
-    for( unsigned int whichPower( 1 );
+    for( size_t whichPower( 1 );
          whichPower < fitCoefficients.size();
          ++whichPower )
     {
       termContribution = fitCoefficients[ whichPower ];
-      for( unsigned int powerCount( 0 );
+      for( size_t powerCount( 0 );
            powerCount < whichPower;
            ++powerCount )
       {
